+++ ukfuelfinder/__init__.py
@@ -4,25 +4,26 @@ UK Fuel Finder Python Library
 Python library for accessing the UK Government Fuel Finder API.
 """
 
-__version__ = "1.2.0"
+__version__ = "2.0.0"
 
 from .client import FuelFinderClient
+from .config import set_global_backward_compatible
 from .exceptions import (
-    FuelFinderError,
-    AuthenticationError,
-    InvalidCredentialsError,
-    TokenExpiredError,
     APIError,
+    AuthenticationError,
+    ConnectionError,
+    FuelFinderError,
+    InvalidCredentialsError,
+    NetworkError,
     NotFoundError,
     RateLimitError,
-    ServerError,
-    ValidationError,
-    NetworkError,
-    TimeoutError,
-    ConnectionError,
     ResponseParseError,
+    ServerError,
+    TimeoutError,
+    TokenExpiredError,
+    ValidationError,
 )
-from .models import PFS, PFSInfo, FuelPrice, Address, Location
+from .models import PFS, Address, FuelPrice, Location, PFSInfo
 
 __all__ = [
     "FuelFinderClient",
@@ -44,4 +45,5 @@ __all__ = [
     "FuelPrice",
     "Address",
     "Location",
+    "set_global_backward_compatible",
 ]